+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 2026-07-30
+- 實作 decode opcode。
 - 實作 VF 暫存器。
 - 判斷 Memory Layout。
 - 重點是做出來，先不要優化，做出來最重要。